refactor(authentication-api): move from main to api package
move interfaces to the api package

related to https://github.com/podman-desktop/podman-desktop/issues/14361
